Majority of the implementation.
I have implemented 4 interfaces and the appropriate implementation
- IResult<TResult, TError> the most generic interface allowing for the largest number of use cases to be achieved.
- IResult<TResult> where TError is IEnumerable<string>
- IVoidResult<TError> TResult is Unit a static class with only instance in existing. Useful for when a value is not returned from a pipeline. All we care about is Success/Failure
- IVoidResult TError is IEnumerable<string>

Still need to decide if the behaviour will be encoded into the interfaces or into extension methods.
I think extension methods are probably are the way forward with the interfaces being the minimal implementation if used just for Switch and Match allowing mocked consumption, eventually phasing out the interfaces all together if achievable.
As this behaviour shouldn't really be mockable, the classes are simple enough and the extension methods are currently tied to the concrete behaviours anyway.
Keeping this behaviour in the commit as a reminder how it worked before changing it.